improvements
- skip decision based on chapter instead of level
- some level numbers changed a bit with the newest build.
- skip over chapters if the detection doesn't work inbetween
- setting: split_chapters (disable to only have one split at the end)
- rename: running => paused (because that's what the memory value of 0/1
indicates)
- don't pause on level value 0 (for brief moments in the loading
screen), because it'll likely cause confusion, as to why the game time
is different from the real time

Likely the level value that I check isn't the real level or scene number
but some kind of resource loaded into the current scene. But it's
somewhat unique for the different chapters, so it can be used to detect
the different chapters. Hopefully its values will not change too often
in the future.